Implemented mouse click to place tiles. You can actually play now!
Change controller to return the game state and an error message,
if a player attempts to claim or place a tile illegally.

Added logic to placing tile that does NOT remove the tile from the
round if the player attempted to place the tile illegally.

Added optional "skipReason" so the backend can report when the last
tile cannot be played. Partially implemented the check.

Javadoc'ed the files that were modified.